+++ Runtime/Extensions/InputDeviceExtension.cs
@@ -23,19 +23,23 @@ namespace ActionCode.InputSystem
 
 #if UNITY_EDITOR || UNITY_ANDROID
             if (device is UnityEngine.InputSystem.Android.AndroidGamepad) type = InputDeviceType.AndroidGamepad;
+            else if (device is UnityEngine.InputSystem.Android.XboxOneGamepadAndroid) type = InputDeviceType.XboxOne;
 #endif
-#if UNITY_EDITOR || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA || UNITY_PS3 || UNITY_PS4 || UNITY_SWITCH
+
+#if UNITY_EDITOR || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA || UNITY_PS3 || UNITY_PS4 || UNITY_PS5 || UNITY_SWITCH
             if (device is UnityEngine.InputSystem.DualShock.DualShock3GamepadHID) type = InputDeviceType.Playstation3;
             else if (device is UnityEngine.InputSystem.DualShock.DualShock4GamepadHID) type = InputDeviceType.Playstation4;
             else if (device is UnityEngine.InputSystem.DualShock.DualSenseGamepadHID) type = InputDeviceType.Playstation5;
             else if (device is UnityEngine.InputSystem.Switch.SwitchProControllerHID) type = InputDeviceType.SwitchProController;
 #endif
+
 #if UNITY_EDITOR || UNITY_IOS || UNITY_TVOS || UNITY_PS4
             if (device is UnityEngine.InputSystem.iOS.DualShock4GampadiOS) type = InputDeviceType.Playstation4;
             if (device is UnityEngine.InputSystem.iOS.DualSenseGampadiOS) type = InputDeviceType.Playstation5;
-            else if (device is UnityEngine.InputSystem.iOS.XboxOneGampadiOS) type = InputDeviceType.XboxSystem;
+            else if (device is UnityEngine.InputSystem.iOS.XboxOneGampadiOS) type = InputDeviceType.XboxOne;
             else if (device is UnityEngine.InputSystem.iOS.iOSGameController) type = InputDeviceType.iOSGamepad;
 #endif
+
 #if UNITY_WEBGL
             if (device is UnityEngine.InputSystem.WebGL.WebGLGamepad)
             {