+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Update Input System package to 1.2.0
-- Detect Playstation5 gamepad
+- Detect Playstation5 and XboxOne gamepad
 
 ## [1.0.0-preview.6] - 2023-06-15
 ### Changed